--Change Log:- -The mouse cursor will be shown in the 3D scenes. If the escape key is pressed, the mouse cursor will be hidden in the FPS camera. -Now there are five different shaders and mesh structs that will represent five different types of meshes that can be rendered. These are ---Mesh::PVertexData which only contains position vertex data. ---Mesh::PUVVertexData which contains position and UV vertex data. ---Mesh::PNVertexData which contains position and normal vertex data. ---Mesh::PUVNVertexData which contains position, UV, and normal vertex data. ---Mesh::PUVNTBVertexData which contains position, UV, normal, tangent, and binormal vertex data.
When converting FBX to BFX, the editor will figure out which vertex version the mesh needs and will pick the correct one and write it to the BFX file. In the engine, the BFXLoader will load the correct vertex struct and shader and renders the mesh accordingly.

The next step would be to sort and render all meshes with similar data structs that use a similer shader to avoid swapping shaders.

-Xinput controller is fully supported now. All that is left is a way to control the vibration.
-Updated the demos.